requirements.txt
Django==4.2.7
psycopg2==2.9.9
python-dotenv==1.0.0
django-allauth==0.58.2
djangorestframework-simplejwt==5.3.0
rest-framework-simplejwt==0.0.2
djangorestframework==3.14.0
dj-rest-auth==5.0.2
django-cors-headers==4.3.1